- ignore workflowitem_document_uploaded from chache when applying event - moved group_query & user_query to domain layer
