+++ api/src/index.ts
@@ -44,6 +44,8 @@ import * as Multichain from "./service";
 import * as Cache from "./service/cache2";
 import StorageServiceClient from "./service/Client_storage_service";
 import * as DocumentValidationService from "./service/document_validation";
+import * as GroupQueryService from "./service/domain/organization/group_query";
+import * as UserQueryService from "./service/domain/organization/user_query";
 import * as GlobalPermissionsGetService from "./service/global_permissions_get";
 import * as GlobalPermissionGrantService from "./service/global_permission_grant";
 import * as GlobalPermissionRevokeService from "./service/global_permission_revoke";
@@ -51,7 +53,6 @@ import * as GroupCreateService from "./service/group_create";
 import * as GroupMemberAddService from "./service/group_member_add";
 import * as GroupMemberRemoveService from "./service/group_member_remove";
 import * as GroupPermissionsListService from "./service/group_permissions_list";
-import * as GroupQueryService from "./service/group_query";
 import * as NotificationListService from "./service/notification_list";
 import * as NotificationMarkReadService from "./service/notification_mark_read";
 import * as ProjectAssignService from "./service/project_assign";
@@ -91,7 +92,6 @@ import * as UserPasswordChangeService from "./service/user_password_change";
 import * as UserPermissionsListService from "./service/user_permissions_list";
 import * as UserPermissionGrantService from "./service/user_permission_grant";
 import * as UserPermissionRevokeService from "./service/user_permission_revoke";
-import * as UserQueryService from "./service/user_query";
 import * as WorkflowitemsReorderService from "./service/workflowitems_reorder";
 import * as WorkflowitemAssignService from "./service/workflowitem_assign";
 import * as WorkflowitemCloseService from "./service/workflowitem_close";
