+++ api/src/lib/validation.ts
@@ -1,7 +1,7 @@
 import { throwParseError } from "../httpd/lib";
 import { ConnToken } from "../service/conn";
 import { ServiceUser } from "../service/domain/organization/service_user";
-import * as UserQuery from "../service/user_query";
+import * as UserQuery from "../service/domain/organization/user_query";
 import { Ctx } from "./ctx";
 
 export function isNonemptyString(x: unknown): boolean {
