+++ api/src/service/document_validation.ts
@@ -1,17 +1,17 @@
 import * as crypto from "crypto";
-import VError = require("verror");
-import { ConnToken } from "./conn";
-import * as Cache from "./cache2";
+import logger from "lib/logger";
 import { Ctx } from "../lib/ctx";
 import * as Result from "../result";
+import * as Cache from "./cache2";
+import { ConnToken } from "./conn";
 import * as DocumentValidate from "./domain/document/document_validate";
+import * as GroupQuery from "./domain/organization/group_query";
 import { ServiceUser } from "./domain/organization/service_user";
 import * as Project from "./domain/workflow/project";
 import * as Subproject from "./domain/workflow/subproject";
 import * as Workflowitem from "./domain/workflow/workflowitem";
-import * as GroupQuery from "./group_query";
 import { store } from "./store";
-import logger from "lib/logger";
+import VError = require("verror");
 
 /**
  * Returns true if the given hash matches the given document.
