+++ api/src/service/domain/organization/group_query.ts
@@ -1,19 +1,16 @@
 import logger from "lib/logger";
 import { VError } from "verror";
-
-import { Ctx } from "../lib/ctx";
-import * as Result from "../result";
-import * as Cache from "./cache2";
-import { ConnToken } from "./conn";
-import { NotFound } from "./domain/errors/not_found";
-import * as Group from "./domain/organization/group";
-import * as GroupGet from "./domain/organization/group_get";
-import { Identity } from "./domain/organization/identity";
-import { ServiceUser } from "./domain/organization/service_user";
-import * as UserRecord from "./domain/organization/user_record";
+import { Ctx } from "../../../lib/ctx";
+import * as Result from "../../../result";
+import { ConnToken } from "../../conn";
+import { NotFound } from "../errors/not_found";
+import * as Cache from "./../../cache2";
+import * as Group from "./group";
+import * as GroupGet from "./group_get";
+import { Identity } from "./identity";
+import { ServiceUser } from "./service_user";
 import { getUser } from "./user_query";
-
-// TODO move groups handling to domain layer + make sure the cache is only refreshed _once_
+import * as UserRecord from "./user_record";
 
 export async function getGroups(
   conn: ConnToken,
