+++ api/src/service/domain/organization/user_query.ts
@@ -1,13 +1,13 @@
 import logger from "lib/logger";
 import { VError } from "verror";
-import { Ctx } from "../lib/ctx";
-import * as Result from "../result";
-import * as Cache from "./cache2";
-import { ConnToken } from "./conn";
-import { NotFound } from "./domain/errors/not_found";
-import { ServiceUser } from "./domain/organization/service_user";
-import * as UserGet from "./domain/organization/user_get";
-import * as UserRecord from "./domain/organization/user_record";
+import { Ctx } from "../../../lib/ctx";
+import * as Result from "../../../result";
+import { ConnToken } from "../../conn";
+import { NotFound } from "../errors/not_found";
+import * as Cache from "./../../cache2";
+import { ServiceUser } from "./service_user";
+import * as UserGet from "./user_get";
+import * as UserRecord from "./user_record";
 
 export async function getUsers(
   conn: ConnToken,
