+++ api/src/service/project_assign.ts
@@ -4,11 +4,11 @@ import { Ctx } from "../lib/ctx";
 import * as Result from "../result";
 import * as Cache from "./cache2";
 import { ConnToken } from "./conn";
+import * as GroupQuery from "./domain/organization/group_query";
 import { Identity } from "./domain/organization/identity";
 import { ServiceUser } from "./domain/organization/service_user";
 import * as Project from "./domain/workflow/project";
 import * as ProjectAssign from "./domain/workflow/project_assign";
-import * as GroupQuery from "./group_query";
 import { store } from "./store";
 
 export async function assignProject(
