+++ api/src/service/project_projected_budget_update.ts
@@ -4,12 +4,12 @@ import { Ctx } from "../lib/ctx";
 import * as Result from "../result";
 import * as Cache from "./cache2";
 import { ConnToken } from "./conn";
+import * as GroupQuery from "./domain/organization/group_query";
 import { ServiceUser } from "./domain/organization/service_user";
 import { CurrencyCode, MoneyAmount } from "./domain/workflow/money";
 import * as Project from "./domain/workflow/project";
 import { ProjectedBudget } from "./domain/workflow/projected_budget";
 import * as ProjectProjectedBudgetUpdate from "./domain/workflow/project_projected_budget_update";
-import * as GroupQuery from "./group_query";
 import { store } from "./store";
 
 export async function updateProjectedBudget(
