+++ api/src/service/project_update.ts
@@ -4,10 +4,10 @@ import { Ctx } from "../lib/ctx";
 import * as Result from "../result";
 import * as Cache from "./cache2";
 import { ConnToken } from "./conn";
+import * as GroupQuery from "./domain/organization/group_query";
 import { ServiceUser } from "./domain/organization/service_user";
 import * as Project from "./domain/workflow/project";
 import * as ProjectUpdate from "./domain/workflow/project_update";
-import * as GroupQuery from "./group_query";
 import { store } from "./store";
 
 export async function updateProject(
