+++ api/src/service/subproject_assign.ts
@@ -4,12 +4,12 @@ import { Ctx } from "../lib/ctx";
 import * as Result from "../result";
 import * as Cache from "./cache2";
 import { ConnToken } from "./conn";
+import * as GroupQuery from "./domain/organization/group_query";
 import { Identity } from "./domain/organization/identity";
 import { ServiceUser } from "./domain/organization/service_user";
 import * as Project from "./domain/workflow/project";
 import * as Subproject from "./domain/workflow/subproject";
 import * as SubprojectAssign from "./domain/workflow/subproject_assign";
-import * as GroupQuery from "./group_query";
 import { store } from "./store";
 
 export async function assignSubproject(
