+++ api/src/service/user_authenticate.ts
@@ -9,15 +9,15 @@ import { getOrganizationAddress } from "../organization/organization";
 import * as Result from "../result";
 import { NotAuthorized } from "./domain/errors/not_authorized";
 import * as AuthToken from "./domain/organization/auth_token";
+import { getGroupsForUser } from "./domain/organization/group_query";
+import * as UserQuery from "./domain/organization/user_query";
 import * as UserRecord from "./domain/organization/user_record";
 import { AuthenticationFailed } from "./errors/authentication_failed";
 import { getselfaddress } from "./getselfaddress";
 import { getGlobalPermissions } from "./global_permissions_get";
 import { grantpermissiontoaddress } from "./grantpermissiontoaddress";
-import { getGroupsForUser } from "./group_query";
 import { importprivkey } from "./importprivkey";
 import { hashPassword, isPasswordMatch } from "./password";
-import * as UserQuery from "./user_query";
 
 export interface UserLoginResponse {
   id: string;
