+++ api/src/service/workflowitem_assign.ts
@@ -5,6 +5,7 @@ import * as Result from "../result";
 import * as Cache from "./cache2";
 import { ConnToken } from "./conn";
 import { BusinessEvent } from "./domain/business_event";
+import * as GroupQuery from "./domain/organization/group_query";
 import { Identity } from "./domain/organization/identity";
 import { ServiceUser } from "./domain/organization/service_user";
 import * as Project from "./domain/workflow/project";
@@ -12,7 +13,6 @@ import * as Subproject from "./domain/workflow/subproject";
 import * as Workflowitem from "./domain/workflow/workflowitem";
 import * as WorkflowitemAssign from "./domain/workflow/workflowitem_assign";
 import * as TypeEvents from "./domain/workflowitem_types/apply_workflowitem_type";
-import * as GroupQuery from "./group_query";
 import { store } from "./store";
 
 export async function assignWorkflowitem(
