+++ api/src/service/workflowitem_close.ts
@@ -5,13 +5,13 @@ import * as Result from "../result";
 import * as Cache from "./cache2";
 import { ConnToken } from "./conn";
 import { BusinessEvent } from "./domain/business_event";
+import * as GroupQuery from "./domain/organization/group_query";
 import { ServiceUser } from "./domain/organization/service_user";
 import * as Project from "./domain/workflow/project";
 import * as Subproject from "./domain/workflow/subproject";
 import * as Workflowitem from "./domain/workflow/workflowitem";
 import * as WorkflowitemClose from "./domain/workflow/workflowitem_close";
 import * as TypeEvents from "./domain/workflowitem_types/apply_workflowitem_type";
-import * as GroupQuery from "./group_query";
 import { store } from "./store";
 
 export async function closeWorkflowitem(
