+++ api/src/service/workflowitem_permission_grant.ts
@@ -8,19 +8,19 @@ import * as PrivateKeyGet from "../organization/organization";
 import * as Result from "../result";
 import * as Cache from "./cache2";
 import { ConnToken } from "./conn";
-import * as DocumentShare from "./domain/document/document_share";
 import * as DocumentGet from "./domain/document/document_get";
+import * as DocumentShare from "./domain/document/document_share";
 import * as SecretGet from "./domain/document/secret_get";
+import * as GroupQuery from "./domain/organization/group_query";
 import { Identity } from "./domain/organization/identity";
 import { ServiceUser } from "./domain/organization/service_user";
+import * as UserQuery from "./domain/organization/user_query";
 import * as Project from "./domain/workflow/project";
 import * as Subproject from "./domain/workflow/subproject";
 import * as Workflowitem from "./domain/workflow/workflowitem";
 import * as WorkflowitemPermissionGrant from "./domain/workflow/workflowitem_permission_grant";
-import * as GroupQuery from "./group_query";
 import * as PublicKeyGet from "./public_key_get";
 import { store } from "./store";
-import * as UserQuery from "./user_query";
 
 export { RequestData } from "./domain/workflow/project_create";
